Move checkBoxInput to component + refactor
