+++ src/components/page/LinkGroupEditor.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import SavableTextInput from "$components/formHelpers/SavableTextInput.svelte";
+  import SavableTextInput from "$components/formHelpers/inputs/SavableTextInput.svelte";
   import { Link, type LinkGroup } from "$lib/domain/Link";
   import { navbarStore } from "$lib/stores/NavbarStore";
   import LinkEditor from "./LinkEditor.svelte";
